particle_dc/requirements.txt
paho-mqtt==1.6.1
chevron==0.14.0
tomli==2.0.1
pyzmq==27.1.0
smbus2
bcr-libraries
spidev
rpi.gpio
sensirion-i2c-sen5x
pyserial